DEATH OF THE PROJECT
As Dgd has released a new song, the cracks in this project are really starting to show. I want to remake everything from scratch and use more object oriented javascript so that I may add more songs instead of hardcoding everything. I might use tables instead of grid because of how flawed grid is when adding more columns etc. For now this is the death of the project, but it shall be revived and reworked soon.